cvat/requirements/base.txt
click==6.7
Django==2.0.3
django-appconf==1.0.2
django-auth-ldap==1.4.0
django-cacheops==4.0.6
django-compressor==2.2
django-rq==1.0.1
EasyProcess==0.2.3
ffmpy==0.2.2
Pillow==5.1.0
numpy==1.14.2
patool==1.12
python-ldap==3.0.0
pytz==2018.3
pyunpack==0.1.2
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
rjsmin==1.0.12
rq==0.10.0
scipy==1.0.1
sqlparse==0.2.4
django-sendfile==0.3.11
dj-pagination==2.3.2
opencv-python==3.4.3.18
thrift==0.11.0
hbase-thrift==0.20.4
google==2.0.1
protobuf==3.6.1
kazoo==2.5.0
lmdb==0.94
happybase==1.1.0
torch==0.4.1
torchvision==0.2.1
matplotlib==3.0.0